app/requirements-gemini-enhanced.txt
# Enhanced Gemini API Dependencies
# Core Google AI library (new version)
google-genai>=0.3.0

# Additional dependencies for enhanced features
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Optional: For advanced features
pillow>=9.0.0  # Image processing
requests>=2.28.0  # HTTP requests
